feat: expose @zwave-js/core/safe
