+++ packages/core/package.json
@@ -7,7 +7,19 @@
   },
   "keywords": [],
   "main": "build/index.js",
+  "exports": {
+    ".": "./build/index.js",
+    "./package.json": "./package.json",
+    "./safe": "./build/index_safe.js"
+  },
   "types": "build/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "build/index_safe.d.ts": [
+        "build/index_safe.d.ts"
+      ]
+    }
+  },
   "files": [
     "build/**/*.{js,d.ts,map}"
   ],
