+++ packages/core/src/index_safe.ts
@@ -1,20 +1,17 @@
+/* @noExternalImports */
+
 export * from "./capabilities/CommandClasses";
 export * from "./capabilities/ControllerCapabilities";
 export * from "./capabilities/NodeInfo";
 export * from "./capabilities/Protocols";
 export * from "./consts";
 export * from "./error/ZWaveError";
-export * from "./log/shared";
 export * from "./log/shared_safe";
 export * from "./security/crypto";
 export * from "./security/DSK";
-export * from "./security/Manager";
-export * from "./security/Manager2";
-export * from "./security/QR";
 export * from "./security/SecurityClass";
 export * from "./test/assertZWaveError";
 export * from "./util/crc";
-export * from "./util/date";
 export * from "./util/firmware";
 export * from "./util/graph";
 export * from "./util/misc";
